requirements.txt
python-dotenv
matplotlib
requests 
